The entire code resides in one `vanity` package
Split it to three separate packages - `core`, `keypair`, `vanity`
